requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
accelerate=0.24.1=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
bcrypt=4.0.1=pypi_0
bitsandbytes=0.37.2=pypi_0
black=23.11.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.16.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.5=pypi_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
datasets=2.14.6=pypi_0
deepspeed=0.8.3=pypi_0
dill=0.3.6=pypi_0
distro=1.8.0=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.1.3=pypi_0
fastapi=0.104.1=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.1=pypi_0
filelock=3.13.1=pypi_0
fire=0.5.0=pypi_0
flit-core=3.9.0=py310h06a4308_0
fonttools=4.44.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gradio=3.9=pypi_0
h11=0.12.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.15.0=pypi_0
httpx=0.25.1=pypi_0
huggingface-hub=0.19.0=pypi_0
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonlines=4.0.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.0.34=0
libcurand=10.3.4.52=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
linkify-it-py=2.0.2=pypi_0
loralib=0.1.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.1=pypi_0
mdit-py-plugins=0.4.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
ninja=1.11.1.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openai=0.27.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
orjson=3.9.10=pypi_0
packaging=23.2=pypi_0
pandas=2.1.2=pypi_0
paramiko=3.3.1=pypi_0
pathspec=0.11.2=pypi_0
peft=0.3.0=pypi_0
pillow=10.1.0=pypi_0
pip=23.3=py310h06a4308_0
platformdirs=4.0.0=pypi_0
protobuf=4.25.0=pypi_0
psutil=5.9.6=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=14.0.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pydantic=1.10.13=pypi_0
pydub=0.25.1=pypi_0
pynacl=1.5.0=pypi_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytorch=1.13.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.31.0=py310h06a4308_0
responses=0.18.0=pypi_0
rouge-score=0.1.2=pypi_0
scipy=1.11.3=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.34.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.1=h5eee18b_0
starlette=0.27.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torchaudio=0.13.1=py310_cu117
torchvision=0.14.1=py310_cu117
tqdm=4.66.1=pypi_0
transformers=4.29.2=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
urllib3=2.0.7=pypi_0
uvicorn=0.24.0.post1=pypi_0
wandb=0.16.0=pypi_0
websockets=12.0=pypi_0
wheel=0.41.2=py310h06a4308_0
xxhash=3.4.1=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
